added missing features

+++ app/src/main/java/com/example/cryptocurrencytracker/utils/graph/DataPoint.kt
@@ -1,6 +1,6 @@
 package com.example.cryptocurrencytracker.utils.graph
 
 data class DataPoint(
-    val xVal: Float,
-    val yVal: Float
+    val xVal: Double,
+    val yVal: Double
 )
